Code cleanup and runtime support (#3)
