+++ Assets/AsepriteImporter.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2e738a64a2c3294299cac20988cd86e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
