+++ Assets/AsepriteImporter/Runtime/Data.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 0a155007e52db4543ace1bd15b376047
-NativeFormatImporter:
+guid: 12e73164bfb7f6e4bbb70f8f48b915dc
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
