+++ Assets/AsepriteImporter/Runtime/Data/AseCel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e3b27ba51141384abb38ff48431d1ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
