+++ Assets/AsepriteImporter/Runtime/Data/AseColor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 834293f1fdbadce488178fa035591ab9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
