+++ Assets/AsepriteImporter/Runtime/Data/AseFrame.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e87ca5463ea8c54fbb5d32d7ba50b3c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
