+++ Assets/AsepriteImporter/Runtime/Data/AseHeader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bfccbc755107694581a3de5058a043e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
