+++ Assets/AsepriteImporter/Runtime/Data/AseLayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fea7f02b1f6d92d4dbb00f13da3cc4e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
