+++ Assets/AsepriteImporter/Runtime/Data/AsePalette.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ab3abbf69d4adc4da2b2d9be22dee4f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
