+++ Assets/AsepriteImporter/Runtime/Data/Aseprite.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad1f78b47fa723348a544457a939b0de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
