+++ Assets/AsepriteImporter/Runtime/Enums.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 648215be9867dee4ca452b3bfe435ca3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
