+++ Assets/AsepriteImporter/Runtime/Enums/BlendMode.cs
@@ -1,0 +1,28 @@
+﻿
+using System;
+
+namespace Assets.AsepriteImporter.Runtime.Enums
+{
+    public enum BlendMode : UInt16
+    {
+        Normal = 0,
+        Multiply = 1,
+        Screen = 2,
+        Overlay = 3,
+        Darken = 4,
+        Lighten = 5,
+        ColorDodge = 6,
+        ColorBurn = 7,
+        HardLight = 8,
+        SoftLight = 9,
+        Difference = 10,
+        Exclusion = 11,
+        Hue = 12,
+        Saturation = 13,
+        Color = 14,
+        Luminosity = 15,
+        Addition = 16,
+        Subtract = 17,
+        Divide = 18,
+    }
+}
