+++ Assets/AsepriteImporter/Runtime/Enums/BlendMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19859fa5815af4f46b64a75c86de3777
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
