+++ Assets/AsepriteImporter/Runtime/Enums/CelType.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Assets.AsepriteImporter.Runtime.Enums
+{
+    public enum CelType : UInt16
+    {
+        Raw = 0,
+        Linked = 1,
+        CompressedImage = 2,
+    }
+}
