+++ Assets/AsepriteImporter/Runtime/Enums/CelType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30434bf54a8073a48b9f5f94c9467ccd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
