+++ Assets/AsepriteImporter/Runtime/Enums/ChunkType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bde6ebf0d4b85aa4a9c96dcd80ebdf02
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
