+++ Assets/AsepriteImporter/Runtime/Enums/ColorDepth.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Assets.AsepriteImporter.Runtime.Enums
+{
+    public enum ColorDepth : UInt16
+    {
+        Undefined = 0,
+        RGBA = 32,
+        Grayscale = 16,
+        Indexed = 8
+    }
+}
