+++ Assets/AsepriteImporter/Runtime/Enums/ColorDepth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53ee93459d05249499a93d1cf9240948
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
