+++ Assets/AsepriteImporter/Runtime/Enums/LayerFlags.cs
@@ -1,0 +1,17 @@
+﻿
+using System;
+
+namespace Assets.AsepriteImporter.Runtime.Enums
+{
+    public enum LayerFlags : UInt16
+    {
+        Undefined = 0,
+        Visible = 1,
+        Editable = 2,
+        LockMovement = 3,
+        Background = 4,
+        PreferLinkedCels = 5,
+        Collapsed = 32,
+        RefLayer = 64
+    }
+}
