+++ Assets/AsepriteImporter/Runtime/Enums/LayerFlags.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c45a3553d778d1041b1ba5b0f3c260ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
