+++ Assets/AsepriteImporter/Runtime/Enums/LayerType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f43b43cf9fac3ba419c9f1e965f10208
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
