--- Assets/Files/Test.aseprite.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0a28be927d1b9b949bd3e5c90340ee63
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 62c9582ebe2158d45ba146092defd5ed, type: 3}
-  Aseprite:
-    _atlas: {instanceID: 0}
